requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer models
timm>=0.9.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Data handling
pandas>=2.0.0

# Model utilities
einops>=0.6.0

# Optional: For training
# tensorboard>=2.12.0
# wandb>=0.15.0
